requirements.txt
gradio==3.36.1
transformers==4.30.2
torch

rss-parser==1.1.0

newspaper3k